deprecated sp objects

These are objects in the older sp format. They have been replaced with simple feature (sf) format and will be eventually removed from the package

Pitch_US_Population_2016_v1

Pitch_US_Population_2016_v1.states

Pitch_US_Population_2016_v1.centers

FiveThirtyEightElectoralCollege

FiveThirtyEightElectoralCollege.states

FiveThirtyEightElectoralCollege.centers

NPR1to1

NPR1to1.centers

DKOS_50_State_OuterHex_Tilemap_v1

DKOS_50_State_InnerHex_Tilemap_v1

DKOS_50_State_Hex_Tilemap_v1.centers

DKOS_CD_Hexmap_v1.1

DKOS_CD_Hexmap_v1.1.states

DKOS_CD_Hexmap_v1.1.centers

DKOS_Electoral_College_Map_v1

DKOS_Electoral_College_Map_v1.states

DKOS_Electoral_College_Map_v1.centers

DKOS_Distorted_Electoral_College_Map_v1

DKOS_Distorted_Electoral_College_Map_v1.centers

NPR.DemersCartogram

NPR.DemersCartogram.centers

WP

WP.centers

WSJ

WSJ.centers

Datamap.io.tilegram

Datamap.io.tilegram.centers

Format

See the specific object documentation for the renamed objects.

Details

Pitch_US_Population_2016_v1 is deprecated; please use sf_Pitch_US_Population_2016_v1 instead.

Pitch_US_Population_2016_v1.states is deprecated; please use sf_Pitch_US_Population_2016_v1.states instead.

Pitch_US_Population_2016_v1.centers is deprecated; please use sf_Pitch_US_Population_2016_v1.centers instead.

FiveThirtyEightElectoralCollege is deprecated; please use sf_FiveThirtyEightElectoralCollege instead.

FiveThirtyEightElectoralCollege.states is deprecated; please use sf_FiveThirtyEightElectoralCollege.states instead.

FiveThirtyEightElectoralCollege.centers is deprecated; please use sf_FiveThirtyEightElectoralCollege.centers instead.

NPR1to1 is deprecated; please use sf_NPR1to1 instead.

NPR1to1.centers is deprecated; please use sf_NPR1to1.centers instead.

DKOS_50_State_OuterHex_Tilemap_v1 is deprecated; please use sf_DKOS_50_State_OuterHex_Tilemap_v1 instead.

DKOS_50_State_InnerHex_Tilemap_v1 is deprecated; please use sf_DKOS_50_State_InnerHex_Tilemap_v1 instead.

DKOS_50_State_Hex_Tilemap_v1.centers is deprecated; please use sf_DKOS_50_State_Hex_Tilemap_v1.centers instead.

DKOS_CD_Hexmap_v1.1 is deprecated; please use sf_DKOS_CD_Hexmap_v1.1 instead.

DKOS_CD_Hexmap_v1.1.states is deprecated; please use sf_DKOS_CD_Hexmap_v1.1.states instead.

DKOS_CD_Hexmap_v1.1.centers is deprecated; please use sf_DKOS_CD_Hexmap_v1.1.centers instead.

DKOS_Electoral_College_Map_v1 is deprecated; please use sf_DKOS_Electoral_College_Map_v1 instead.

DKOS_Electoral_College_Map_v1.states is deprecated; please use sf_DKOS_Electoral_College_Map_v1.states instead.

DKOS_Electoral_College_Map_v1.centers is deprecated; please use sf_DKOS_Electoral_College_Map_v1.centers instead.

DKOS_Distorted_Electoral_College_Map_v1 is deprecated; please use sf_DKOS_Distorted_Electoral_College_Map_v1 instead.

DKOS_Distorted_Electoral_College_Map_v1.centers is deprecated; please use sf_DKOS_Distorted_Electoral_College_Map_v1.centers instead.

NPR.DemersCartogram is deprecated; please use sf_NPR.DemersCartogram instead.

NPR.DemersCartogram.centers is deprecated; please use sf_NPR.DemersCartogram.centers instead.

WP is deprecated; please use sf_WP instead.

WP.centers is deprecated; please use sf_WP.centers instead.

WSJ is deprecated; please use sf_WSJ instead.

WSJ.centers is deprecated; please use sf_WSJ.centers instead.

Datamap.io.tilegram is deprecated; please use sf_Datamap.io.tilegram instead.

Datamap.io.tilegram.centers is deprecated; please use sf_Datamap.io.tilegram.centers instead.